Adionada as validações e feita uma atualização no método "UpdateAsync()" do controlador "VeiculoController";  também foi atualizada a assinatura do método "UpdateAsync()" da interface "IVeiculoRepository".

+++ GerenciamentoFrotaVeiculo.Api/Controllers/VeiculosController.cs
@@ -20,13 +20,23 @@ namespace GerenciamentoFrotaVeiculo.Controllers
         {
             var veiculo = await _veiculoRepository.GetAsync(id);
 
+            if (veiculo is null)
+            {
+                return NotFound();
+            }
+
             return Ok(veiculo);
         }
 
-        [HttpGet("veiculos/")]
+        [HttpGet("veiculos")]
         public async Task<IActionResult> GetAllAsync()
         {
             var veiculos = await _veiculoRepository.GetAllAsync();
+
+            if (veiculos.Count == 0)
+            {
+                return NotFound();
+            }
 
             return Ok(veiculos);
         }
@@ -34,23 +44,51 @@ namespace GerenciamentoFrotaVeiculo.Controllers
         [HttpPost("veiculos")]
         public async Task<IActionResult> CreateAsync([FromBody] Veiculo veiculo)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
+
             await _veiculoRepository.CreateAsync(veiculo);
 
             return Ok(veiculo);
         }
 
         [HttpPut("veiculos/{id}")]
-        public async Task<IActionResult> UpdateAsync([FromBody] Veiculo veiculo)
+        public async Task<IActionResult> UpdateAsync([FromBody] Veiculo veiculoRequisicao, int id)
         {
-            await _veiculoRepository.UpdateAsync(veiculo);
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
 
-            return Ok(veiculo);
+            if (veiculoRequisicao.Id != id)
+            {
+                return BadRequest();
+            }
+
+            var veiculoDb = await _veiculoRepository.GetAsync(id);
+
+            if (veiculoDb is null)
+            {
+                return NotFound();
+            }
+
+            await _veiculoRepository.UpdateAsync(veiculoDb, veiculoRequisicao);
+
+            return Ok(veiculoRequisicao);
         }
 
         [HttpDelete("veiculos/{id}")]
         public async Task<IActionResult> DeleteAsync(int id)
         {
             var veiculo = await _veiculoRepository.GetAsync(id);
+
+            if (veiculo is null)
+            {
+                return NotFound();
+            }
+
             await _veiculoRepository.DeleteAsync(veiculo);
 
             return Ok(veiculo);
